requirement.txt
Flask==3.1.1
torch==2.7.1
torchvision==0.22.1
Pillow==11.2.1
